% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{abundance_jaccard}
\alias{abundance_jaccard}
\alias{abundance_sorenson}
\title{Chao's abundance-weighted indices}
\usage{
abundance_jaccard(x, y)

abundance_sorenson(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\value{
The dissimilarity between \code{x} and \code{y}
}
\description{
These indices were first developed and introduced as probabilistic,
abundance-weighted versions of the Jaccard and Sorenson indices. The
abundance information is summarized by two quantities, named \eqn{U} and
\eqn{V} in the literature. Each quantity indicates the total abundance of
species that are common to both samples: \eqn{U} is the total abundance of
such species in the first sample, and \eqn{V} the total in the second sample.
}
\details{
In addition to defining the quantities \eqn{U} and \eqn{V}, the articles
provide formulas for estimating the value \eqn{U} and \eqn{V} if more
sampling was to be performed. In other software packages where these indices
are implemented, it is the estimated quantities, \eqn{U_{est}} and
\eqn{V_{est}} that are used to calculate the distance. For example, the
\code{chaodist} function in Vegan uses the estimators. In this package, we
use the observed values of \eqn{U} and \eqn{V}, rather than the estimated
values. Thus, the distances returned here may differ from other software
packages that use the estimators rather than the observed values of \eqn{U}
and \eqn{V}. In general, the estimators increase \eqn{U} and \eqn{V},
resulting in greater estimated abundance of shared species and thus a
smaller distance between samples.

The \emph{abundance-weighted Jaccard} distance is \eqn{UV / (U + V - UV)}.
Relation of \code{abundance_jaccard()} to other definitions:
\itemize{
  \item Differs \code{vegdist()} with \code{method = "chao"}. We use the
    observed values, not the estimators.
  \item Differs from the \code{jabund} calculator in Mothur. We use the
    observed values, not the estimators.
}

The \emph{abundance-weighted Sorenson} distance is \eqn{2UV / (U + V)}.
Relation of \code{abundance_sorenson()} to other definitions:
\itemize{
  \item Differs from the \code{sabund} calculator in Mothur. We use the
    observed values, not the estimators.
}
}
\references{
Chao A, Chazdon RL, Colwell R, Shen TJ. A new statistical approach for
assessing similarity of species composition with incidence and abundance
data. Ecology Letters 2005;8:148-159.
}
