% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\docType{data}
\name{alpha_diversities}
\alias{alpha_diversities}
\title{All alpha diversity measures}
\format{An object of class \code{character} of length 16.}
\usage{
alpha_diversities
}
\description{
We exclude functions that return multiple values.
}
\keyword{datasets}
