% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\docType{data}
\name{beta_diversities}
\alias{beta_diversities}
\title{All beta-diversity measures}
\format{An object of class \code{character} of length 32.}
\usage{
beta_diversities
}
\description{
We exclude the phylogenetic measures and functions that return multiple
values.
}
\keyword{datasets}
