% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{bray_curtis}
\alias{bray_curtis}
\title{Bray-Curtis distance}
\usage{
bray_curtis(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
The Bray-Curtis distance is the Manhattan distance divided by the sum of
both vectors.
}
\details{
For two vectors \code{x} and \code{y}, the Bray-Curtis distance is defined
as \deqn{d(x, y) = \frac{\sum_i |x_i - y_i|}{\sum_i x_i + y_i}}.
Relation to other definitions:
\itemize{
  \item Equivalent to \code{vegdist()} with \code{method = "bray"}.
  \item Equivalent to \eqn{D_{14} = 1 - S_{17}}{D_14 = 1 - S_17} in
    Legendre & Legendre.
}
}
