% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_components.R
\name{bray_curtis_balanced_component}
\alias{bray_curtis_balanced_component}
\alias{bray_curtis_gradient_component}
\alias{ruzicka_balanced_component}
\alias{ruzicka_gradient_component}
\title{Balanced variation and abundance gradient components for abundance data}
\usage{
bray_curtis_balanced_component(x, y)

bray_curtis_gradient_component(x, y)

ruzicka_balanced_component(x, y)

ruzicka_gradient_component(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Balanced variation and abundance gradient components for abundance data
}
\references{
Baselga A. Separating the two components of abundance-based dissimilarity:
balanced changes in abundance vs.abundance gradients. Methods in Ecology
and Evolution. 2013;4:552–557.

Baselga A. Partitioning abundance-based multiple-site dissimilarity into
components: balanced variation in abundance and abundance gradients.
Methods in Ecology and Evolution. 2017;8:799–808.
}
