% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{canberra}
\alias{canberra}
\title{Canberra distance}
\usage{
canberra(x, y)
}
\arguments{
\item{x, y}{Numeric vectors

Equivalent to R's built-in dist() function with method = "canberra".
Equivalent to vegdist() with method = "canberra", multiplied by the number of nonzero entries.
Equivalent to canberra() function in scipy.spatial.distance.
Equivalent to D_10 in Legendre & Legendre.}
}
\description{
Canberra distance
}
