% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{cao}
\alias{cao}
\title{Cao index of dissimilarity}
\usage{
cao(x, y)
}
\description{
Cao index of dissimilarity
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to vegdist() with method = "cao".
}
}
