% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{chebyshev}
\alias{chebyshev}
\title{Chebyshev distance}
\usage{
chebyshev(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Chebyshev distance
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to chebyshev() function in scipy.spatial.distance.
}
}
