% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{chord}
\alias{chord}
\title{Chord distance}
\usage{
chord(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
The Euclidean distance after scaling each community vector by the root sum
of squares. Ranges from 0 to \eqn{\sqrt{2}}{sqrt(2)}.
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to \eqn{D_3}{D_3} in Legendre & Legendre.
}
}
