% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{clark_coefficient_of_divergence}
\alias{clark_coefficient_of_divergence}
\title{Clark's coefficient of divergence}
\usage{
clark_coefficient_of_divergence(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Clark's coefficient of divergence
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to \eqn{D_{11}}{D_11} in Legendre & Legendre.
}
}
