% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{cy_dissimilarity}
\alias{cy_dissimilarity}
\title{CY index of dissimilarity}
\usage{
cy_dissimilarity(x, y, base = 10, min_value = 0.1)
}
\arguments{
\item{x, y}{Numeric vectors}

\item{base}{Base of the logarithm}

\item{min_value}{Replacement for zero or near-zero values}
}
\description{
The CY or Cao index of dissimilarity
}
\details{
The CY index was designed to be used with whole-numbered counts, rather than
proportions. It is unbounded, unlike the Bray-Curtis distance, which has an
upper bound of 1. The CY index is defined as
\deqn{
  d(x,y) = \frac{1}{N} \sum_i
  \left (
    \frac{
      (x_i + y_i) \log_{10} ( \frac{x_i + y_i}{2} ) +
      x_i \log_{10}(y_i) + y_i \log_{10}(x_i)
    }{
      x_i + y_i
    }
  \right ),
}
where \eqn{N} is the total number of species in vectors \eqn{x} and \eqn{y}.
Double zeros are not considered in the measure. When either \eqn{x_i} or
\eqn{y_i} are zero, they need to be replaced by another value to avoid
infinite values in the sum. Cao suggested replacing zero values with
\eqn{0.1}, which is one log lower than the minimum value for whole-numbered
counts.

It probably doesn't make sense to use this index with species proportions,
since the measure was created to measure log-scale differences in species
counts.

Relation to other definitions:
\itemize{
  \item Equivalent to the vegdist() function with method = "cao", if
  \code{base = exp(1)}.
}
}
\references{
Cao Y, Williams WP, Bark AW. Similarity measure bias in river benthic
Aufwuchs community analysis. Water Environment Research 1997;69(1):95-106.
}
