% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\docType{data}
\name{ecofuncs_alpha}
\alias{ecofuncs_alpha}
\title{All alpha diversity measures}
\format{An object of class \code{character} of length 23.}
\usage{
ecofuncs_alpha
}
\description{
We exclude functions that return multiple values.
}
\keyword{datasets}
