% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifrac.R
\docType{data}
\name{faith_tree}
\alias{faith_tree}
\title{Example data for Faith's phylogenetic diversity}
\format{\code{faith_tree} is a phylogenetic tree with five tips, labeled
a-e. It was created with the \code{ape} library.}
\source{
Faith DP, Richards ZT. Biology (Basel). 2012 Dec 14;1(3):906-32.
10.3390/biology1030906
}
\usage{
faith_tree
}
\description{
This example was used to illustrate phylogenetic diversity in Faith and
Richards (2012).
}
\keyword{datasets}
