% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifrac.R
\name{generalized_unifrac}
\alias{generalized_unifrac}
\title{Generalized UniFrac distance}
\usage{
generalized_unifrac(x, y, tree, alpha = 0.5)
}
\description{
Generalized UniFrac distance
}
\references{
Chen et al. Bioinformatics. 2012;28(16):2106-13.
}
