% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{geodesic_metric}
\alias{geodesic_metric}
\title{Geodesic metric}
\usage{
geodesic_metric(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
If the communities are arranged into vectors of unit length, the geodesic
metric is the length of the arc on a hypersphere between the vectors.
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to \eqn{D_4}{D_4} in Legendre & Legendre.
}
}
