% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{geodisc_metric}
\alias{geodisc_metric}
\title{Geodisc metric}
\usage{
geodisc_metric(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Geodisc metric
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to D_4 in Legendre & Legendre.
}
}
