% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{goods_coverage}
\alias{goods_coverage}
\title{Good's coverage estimator}
\usage{
goods_coverage(x)
}
\description{
Good's coverage estimator
}
\details{
For a vector of species counts \code{x}, Good's coverage estimator is
\eqn{1 - n_1 / N}, where \eqn{n_1} is the number of species with a count
value of 1, and \eqn{N} is the total number of counts in the vector. This
function makes sense only for raw count data, not proportions.

Equivalent to \code{goods_coverage()} in \code{skbio.diversity.alpha}.
}
\examples{
x <- c(9, 0, 1, 2, 5, 2, 1, 1, 0, 7, 2, 1, 0, 1, 1)
goods_coverage(x) # 1 - 6 / 33
goods_coverage(c(5, 4, 3, 2, 1)) # 1 - 1 / 15
}
\references{
Good IJ. The Population Frequencies of Species and the Estimation of
Population Parameters. Biometrika. 1953;40:237-264.
}
