% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{gower}
\alias{gower}
\title{Gower distance}
\usage{
gower(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Gower distance
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to vegdist() with method = "gower".
}
}
