% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{hamming}
\alias{hamming}
\title{Hamming distance}
\usage{
hamming(x, y)
}
\description{
Equivalent to hamming() function in scipy.spatial.distance
For binary data, equivalent to 1 - S_1 in Legendre & Legendre.
}
