% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{hellinger}
\alias{hellinger}
\title{Hellinger distance}
\usage{
hellinger(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
The Hellinger distance is the chord distance computed on species proportions
after a square-root transformation.
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to \eqn{D_{17}}{D_17} in Legendre & Legendre.
}
}
