% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard index of dissimilarity}
\usage{
jaccard(x, y)
}
\description{
Jaccard index of dissimilarity
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to R's built-in \code{dist()} function with
    \code{method = "binary"}.
  \item Equivalent to \code{vegdist()} with \code{method = "jaccard"}
    and \code{binary = TRUE}.
  \item Equivalent to the \code{jaccard()} function in
    \code{scipy.spatial.distance}.
  \item Equivalent to \eqn{1 - S_7}{1 - S_7} in Legendre & Legendre.
}
}
