% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{kulczynski}
\alias{kulczynski}
\title{Kulczynski distance}
\usage{
kulczynski(x, y)
}
\arguments{
\item{x, y}{Numeric vectors

Equivalent to vegdist() with method = "kulczynski".}
}
\description{
Kulczynski distance
}
