% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{kulczynski_cody}
\alias{kulczynski_cody}
\title{Kulczynski-Cody distance}
\usage{
kulczynski_cody(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Kulczynski-Cody distance
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to \eqn{1 - S_{13}}{1 - S_13} in Legendre & Legendre.
}
}
