% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{kulczynski_mothur}
\alias{kulczynski_mothur}
\title{Kulczynski distance (Mothur implementation)}
\usage{
kulczynski_mothur(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Kulczynski distance (Mothur implementation)
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to 1 - S_12 in Legendre & Legendre.
}
}
