% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_components.R
\docType{data}
\name{leprieur_tree}
\alias{leprieur_tree}
\title{Example data for phylogenetic nestedness and turnover components}
\format{
\code{leprieur_tree} is a phylogenetic tree with 8 tips, labeled
a-h. It was created with the \code{ape} library. All edges (branches) in the
tree are of length 1.
}
\source{
Leprieur F, Albouy C, De Bortoli J, Cowman PF, Bellwood DR, Mouillot D.
Quantifying phylogenetic beta diversity: distinguishing between "true"
turnover of lineages and phylogenetic diversity gradients. PLoS One.
2012;7(8):e42760. 10.1371/journal.pone.0042760
}
\usage{
leprieur_tree
}
\description{
This tree was used in Figure 2 of Leprieur et al. (2005) to demonstrate
the nestedness and turnover components of phylogenetic beta diversity.
}
\keyword{datasets}
