% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{manhattan}
\alias{manhattan}
\title{Manhattan or city block distance}
\usage{
manhattan(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
The Manhattan distance is the sum of absolute differences between the
elements of two vectors.
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to R's built-in \code{dist()} function with
    \code{method = "manhattan"}.
  \item Equivalent to \code{vegdist()} with \code{method = "manhattan"}.
  \item Equivalent to the \code{cityblock()} function in
    \code{scipy.spatial.distance}.
  \item Equivalent to \eqn{D_7}{D_7} in Legendre & Legendre.
  \item Whittaker's index of assiciation (\eqn{D_9}{D_9} in Legendre &
    Legendre) is the Manhattan distance computed after transforming to
    proportions and dividing by 2.
}
}
