% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{millar}
\alias{millar}
\title{Binomial index of dissimilarity}
\usage{
millar(x, y)
}
\description{
Binomial index of dissimilarity
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to vegdist() with method = "binomial".
}
}
