% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{modified_mean_character_difference}
\alias{modified_mean_character_difference}
\title{Modified mean character difference}
\usage{
modified_mean_character_difference(x, y)
}
\description{
Equivalent to D_19 in Legendre & Legendre.
For binary data, equivalent to Jaccard distance.
}
