% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{morisita}
\alias{morisita}
\title{Morisita index of dissimilarity}
\usage{
morisita(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Morisita index of dissimilarity
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to vegdist() with method = "morisita".
}
}
