% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{pivot_to_numeric_matrix}
\alias{pivot_to_numeric_matrix}
\title{Convert a data frame in long format to a numeric matrix}
\usage{
pivot_to_numeric_matrix(data, obs_col, feature_col, value_col)
}
\description{
Convert a data frame in long format to a numeric matrix
}
