% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{rms_distance}
\alias{rms_distance}
\title{Root-mean-square distance}
\usage{
rms_distance(x, y)
}
\arguments{
\item{x, y}{Numeric vectors

Equivalent to D_XXXX in Legendre & Legendre.}
}
\description{
Root-mean-square distance
}
