% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{robbins}
\alias{robbins}
\title{Robbins' estimator for the probability of unobserved outcomes}
\usage{
robbins(x)
}
\arguments{
\item{x}{A numeric vector of raw species counts.}
}
\description{
Robbins' estimator for the probability of unobserved outcomes
}
\details{
Robbins' estimator is \eqn{F_1 / (N + 1)}, where \eqn{F_1} is the number of
species observed only once, and \eqn{N} is the total number of counts.
Relation to other definitions:
\itemize{
  \item Equivalent to \code{robbins()} in \code{skbio.diversity.alpha}.
  \item Equivalent to \code{1 - goods_estimator(x)}.
}
}
\examples{
x <- c(9, 0, 1, 2, 5, 2, 1, 1, 0, 7, 2, 1, 0, 1, 1)
robbins(x) # 6 / 33
}
