% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{russel_rao}
\alias{russel_rao}
\title{Russel-Rao distance}
\usage{
russel_rao(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Russel-Rao distance
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to the \code{russelrao()} function in
    \code{scipy.spatial.distance}.
  \item Equivalent to \code{1 - S_{11}}{1 - S_11} in Legendre & Legendre.
}
}
