% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{sokal_michener}
\alias{sokal_michener}
\title{Sokal-Michener distance}
\usage{
sokal_michener(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Sokal-Michener distance
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to... XXXXX
}
}
