% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{sokal_sneath}
\alias{sokal_sneath}
\title{Sokal-Sneath distance}
\usage{
sokal_sneath(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Sokal-Sneath distance
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to sokalsneath() function in scipy.spatial.distance.
  \item Equivalent to 1 - S_10 in Legendre & Legendre.
}
}
