% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{sorenson}
\alias{sorenson}
\title{Sorenson or Dice index of dissimilarity}
\usage{
sorenson(x, y)
}
\description{
Sorenson or Dice index of dissimilarity
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to dice() function in scipy.spatial.distance.
  \item Equivalent to D_13 = 1 - S_8 in Legendre & Legendre.
}
}
