% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{spread_to_numeric_matrix}
\alias{spread_to_numeric_matrix}
\title{Convert a data frame in long format to a numeric matrix}
\usage{
spread_to_numeric_matrix(data, row_key, col_key, value)
}
\description{
Convert a data frame in long format to a numeric matrix
}
