% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifrac.R
\name{unifrac}
\alias{unifrac}
\alias{unweighted_unifrac}
\alias{weighted_normalized_unifrac}
\alias{weighted_unifrac}
\alias{generalized_unifrac}
\alias{variance_adjusted_unifrac}
\alias{information_unifrac}
\title{UniFrac distance}
\usage{
unweighted_unifrac(x, y, tree, xy_labels = NULL)

weighted_normalized_unifrac(x, y, tree, xy_labels = NULL)

weighted_unifrac(x, y, tree, xy_labels = NULL)

generalized_unifrac(x, y, tree, alpha = 0.5, xy_labels = NULL)

variance_adjusted_unifrac(x, y, tree, xy_labels = NULL)

information_unifrac(x, y, tree, xy_labels = NULL)
}
\arguments{
\item{x, y}{Numeric vectors of species counts or proportions.}

\item{tree}{A phylogenetic tree object.}

\item{xy_labels}{A character vector of species labels for \code{x} and
\code{y}.}
}
\description{
UniFrac distance
}
\details{
Information UniFrac distance from "Expanding the UniFrac Toolbox"
}
\references{
Chang et al. BMC Bioinformatics 2011, 12:118.

Chen et al. Bioinformatics. 2012;28(16):2106-13.
}
