% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifrac.R
\name{variance_adjusted_unifrac}
\alias{variance_adjusted_unifrac}
\title{Variance-adjusted UniFrac distance}
\usage{
variance_adjusted_unifrac(x, y, tree, alpha = 0.5, xy_labels = NULL)
}
\description{
Variance-adjusted UniFrac distance
}
\references{
Chang et al. BMC Bioinformatics 2011, 12:118.
}
