% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifrac.R
\name{weighted_normalized_unifrac}
\alias{weighted_normalized_unifrac}
\title{Weighted normalized UniFrac distance}
\usage{
weighted_normalized_unifrac(x, y, tree)
}
\description{
Weighted normalized UniFrac distance
}
