% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{yule}
\alias{yule}
\title{Yule distance}
\usage{
yule(x, y)
}
\description{
Equivalent to yule() function in scipy.spatial.distance.
Equivalent to 1 - S, where S is the Yule coefficient in Legendre & Legendre.
}
