% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{alt_gower}
\alias{alt_gower}
\title{Alternate Gower distance}
\usage{
alt_gower(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Alternate Gower distance
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to vegdist() with method = "altGower".
}
}
