% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{bray_curtis}
\alias{bray_curtis}
\title{Bray-Curtis distance}
\usage{
bray_curtis(x, y)
}
\arguments{
\item{x, y}{Numeric vectors

Equivalent to vegdist() with method = "bray".
Equivalent to D_14 = 1 - s_17 in Legendre & Legendre.}
}
\description{
Bray-Curtis distance
}
