% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{bray_curtis}
\alias{bray_curtis}
\title{Bray-Curtis distance}
\usage{
bray_curtis(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
The Bray-Curtis distance is the sum of absolute differences between the
vectors divided by the sum of both vectors.
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to \code{vegdist()} with \code{method = "bray"}.
  \item Equivalent to \eqn{D_{14} = 1 - S_{17}}{D_14 = 1 - S_17} in
    Legendre & Legendre.
}
}
