% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{canberra}
\alias{canberra}
\title{Canberra distance}
\usage{
canberra(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Canberra distance
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to R's built-in dist() function with method = "canberra".
  \item Equivalent to vegdist() with method = "canberra", multiplied by the number of nonzero entries.
  \item Equivalent to canberra() function in scipy.spatial.distance.
  \item Equivalent to D_10 in Legendre & Legendre.
}
}
