% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{canberra}
\alias{canberra}
\title{Canberra distance}
\usage{
canberra(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
The sum over all elements of the absolute difference in abundance divided by
the sum.
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to R's built-in \code{dist()} function with
    \code{method = "canberra"}.
  \item Equivalent to the \code{vegdist()} function with
    \code{method = "canberra"}, multiplied by the number of nonzero entries.
  \item Equivalent to the \code{canberra()} function in
    \code{scipy.spatial.distance}.
  \item Equivalent to \eqn{D_{10}}{D_10} in Legendre & Legendre.
}
}
