% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{chebyshev}
\alias{chebyshev}
\title{Chebyshev distance}
\usage{
chebyshev(x, y)
}
\arguments{
\item{x, y}{Numeric vectors

Equivalent to chebyshev() function in scipy.spatial.distance.}
}
\description{
Chebyshev distance
}
