% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{chord}
\alias{chord}
\title{Chord distance}
\usage{
chord(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Chord distance
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to D_3 in Legendre & Legendre.
}
}
