% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{clark_coefficient_of_divergence}
\alias{clark_coefficient_of_divergence}
\title{Clark's coefficient of divergence}
\usage{
clark_coefficient_of_divergence(x, y)
}
\description{
Equivalent to D_11 in Legendre & Legendre.
}
