% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{correlation}
\alias{correlation}
\title{Correlation distance}
\usage{
correlation(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Correlation distance
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to correlation() function in scipy.spatial.distance.
}
}
