% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean distance}
\usage{
euclidean(x, y)
}
\arguments{
\item{x, y}{Numeric vectors

Equivalent to R's built-in dist() function with method = "euclidean".
Equivalent to vegdist() with method = "euclidean".
Equivalent to euclidean() function in scipy.spatial.distance.
Equivalent to D_1 in Legendre & Legendre.
Equivalent to D_18 in Legendre & Legendre after transformation to relative abundance.}
}
\description{
Euclidean distance
}
