% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean distance}
\usage{
euclidean(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Euclidean distance
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to R's built-in \code{dist()} function with
    \code{method = "euclidean"}.
  \item Equivalent to \code{vegdist()} with \code{method = "euclidean"}.
  \item Equivalent to the \code{euclidean()} function in
    \code{scipy.spatial.distance}.
  \item Equivalent to \eqn{D_1}{D_1} in Legendre & Legendre.
  \item Equivalent to \eqn{D_{18}}{D_18} in Legendre & Legendre after
    transformation to relative abundance.
}
}
