% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{hamming}
\alias{hamming}
\title{Hamming distance}
\usage{
hamming(x, y)
}
\description{
Hamming distance
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to the \code{hamming()} function in
    \code{scipy.spatial.distance}.
  \item For binary data, equivalent to \code{1 - S_1}{1 - S_1} in Legendre
    & Legendre.
}
}
