% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{hellinger}
\alias{hellinger}
\title{Hellinger distance}
\usage{
hellinger(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Hellinger distance
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to D_17 in Legendre & Legendre.
}
}
