% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard index of dissimilarity}
\usage{
jaccard(x, y)
}
\description{
Jaccard index of dissimilarity
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to R's built-in dist() function with method = "binary".
  \item Equivalent to vegdist() with method = "jaccard" and binary = TRUE.
  \item Equivalent to jaccard() function in scipy.spatial.distance.
  \item Equivalent to 1 - S7 in Legendre & Legendre.
}
}
