% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard index of dissimilarity}
\usage{
jaccard(x, y)
}
\description{
Equivalent to R's built-in dist() function with method = "binary".
Equivalent to vegdist() with method = "jaccard" and binary = TRUE.
Equivalent to jaccard() function in scipy.spatial.distance.
Equivalent to 1 - S7 in Legendre & Legendre.
}
