% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{manhattan}
\alias{manhattan}
\title{Manhattan or city block distance}
\usage{
manhattan(x, y)
}
\description{
Equivalent to R's built-in dist() function with method = "manhattan".
Equivalent to vegdist() with method = "manhattan".
Equivalent to cityblock() function in scipy.spatial.distance.
Equivalent to D_7 in Legendre & Legendre.
Whittaker's index of assiciation (D_9 in Legendre & Legendre) is the
Manhattan distance computed after transforming to proportions and dividing
by 2.
}
