% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{mean_character_difference}
\alias{mean_character_difference}
\title{Mean character difference}
\usage{
mean_character_difference(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
The mean character difference is the Manhattan distance divided by the
vector length.
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to \eqn{D_8}{D_8} in Legendre & Legendre.
  \item For binary data, equivalent to \eqn{1 - S_1}{1 - S_1} in Legendre
    & Legendre, where \eqn{S_1}{S_1} is the simple matching coefficient.
}
}
