% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{mean_character_difference}
\alias{mean_character_difference}
\title{Mean character difference}
\usage{
mean_character_difference(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Mean character difference
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to D_8 in Legendre & Legendre.
  \item For binary data, equivalent to 1 - S_1 in Legendre & Legendre.
}
}
