% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{mean_character_difference}
\alias{mean_character_difference}
\title{Mean character difference}
\usage{
mean_character_difference(x, y)
}
\description{
Equivalent to D_8 in Legendre & Legendre.
For binary data, equivalent to 1 - S_1 in Legendre & Legendre.
}
