% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{minkowski}
\alias{minkowski}
\title{Minkowski distance}
\usage{
minkowski(x, y, p)
}
\description{
Equivalent to R's built-in dist() function with method = "minkowski".
Equivalent to minkowski() function in scipy.spatial.distance.
Equivalent to D_6 in Legendre & Legendre.
}
