% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{minkowski}
\alias{minkowski}
\title{Minkowski distance}
\usage{
minkowski(x, y, p = 1)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Minkowski distance
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to R's built-in dist() function with method = "minkowski".
  \item Equivalent to minkowski() function in scipy.spatial.distance.
  \item Equivalent to D_6 in Legendre & Legendre.
}
}
