% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{modified_mean_character_difference}
\alias{modified_mean_character_difference}
\title{Modified mean character difference}
\usage{
modified_mean_character_difference(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
The modified mean character difference is the Manhattan distance divided by
the number of elements where the vectors are not both equal to zero.
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to \eqn{D_{19}}{D_19} in Legendre & Legendre.
  \item For binary data, equivalent to the Jaccard distance.
}
}
