% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{modified_mean_character_difference}
\alias{modified_mean_character_difference}
\title{Modified mean character difference}
\usage{
modified_mean_character_difference(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Modified mean character difference
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to D_19 in Legendre & Legendre.
  \item For binary data, equivalent to Jaccard distance.
}
}
