% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{morisita}
\alias{morisita}
\title{Morisita index of dissimilarity}
\usage{
morisita(x, y)
}
\arguments{
\item{x, y}{Numeric vectors

Equivalent to vegdist() with method = "morisita".}
}
\description{
Morisita index of dissimilarity
}
