% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{rogers_tanimoto}
\alias{rogers_tanimoto}
\title{Rogers-Tanimoto distance}
\usage{
rogers_tanimoto(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Rogers-Tanimoto distance
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to the \code{rogerstanimoto()} function in
    \code{scipy.spatial.distance}.
  \item Equivalent to \eqn{1 - S_2}{1 - S_2} in Legendre & Legendre.
}
}
