% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{rogers_tanimoto}
\alias{rogers_tanimoto}
\title{Rogers-Tanimoto distance}
\usage{
rogers_tanimoto(x, y)
}
\arguments{
\item{x, y}{Numeric vectors

Equivalent to rogerstanimoto() function in scipy.spatial.distance.
Equivalent to 1 - S_2 in Legendre & Legendre}
}
\description{
Rogers-Tanimoto distance
}
