% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{russel_rao}
\alias{russel_rao}
\title{Russel-Rao distance}
\usage{
russel_rao(x, y)
}
\description{
Equivalent to russelrao() function in scipy.spatial.distance.
Equivalent to 1 - S_11 in Legendre & Legendre
}
