% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{sorenson}
\alias{sorenson}
\title{Sorenson or Dice index of dissimilarity}
\usage{
sorenson(x, y)
}
\description{
After transforming to presence/absence, the average fraction of species
found in one sample but not the other.
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to the \code{dice()} function in
    \code{scipy.spatial.distance}.
  \item Equivalent to \eqn{D_{13} = 1 - S_8}{D_13 = 1 - S_8} in Legendre &
    Legendre.
}
}
