% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_components.R
\name{unifrac_components}
\alias{unifrac_components}
\alias{unweighted_unifrac_turnover_component}
\alias{unweighted_unifrac_nestedness_component}
\title{Nestedness and turnover components of unweighted UniFrac distance}
\usage{
unweighted_unifrac_turnover_component(x, y, tree, xy_labels = NULL)

unweighted_unifrac_nestedness_component(x, y, tree, xy_labels = NULL)
}
\arguments{
\item{x, y}{Numeric vectors of species counts or proportions.}

\item{tree}{A phylogenetic tree object.}

\item{xy_labels}{A character vector of species labels for \code{x} and
\code{y}.}
}
\value{
The nestedness or turnover component of the UniFrac distance
  between communities \code{x} and \code{y}.
}
\description{
Nestedness and turnover components of unweighted UniFrac distance
}
\details{
Leprieur et al. (2012) showed that measures of phylogenetic beta diversity
could be partitioned into nestedness and turnover components, following the
approach of Baselga (2010) for Sorenson dissimilarity.
}
\references{
Leprieur F, Albouy C, De Bortoli J, Cowman PF, Bellwood DR, Mouillot D.
Quantifying phylogenetic beta diversity: distinguishing between "true"
turnover of lineages and phylogenetic diversity gradients. PLoS One.
2012;7(8):e42760. 10.1371/journal.pone.0042760
}
