% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{yule}
\alias{yule}
\title{Yule distance}
\usage{
yule(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Yule distance
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to the \code{yule()} function in
    \code{scipy.spatial.distance}.
  \item Equivalent to \eqn{1 - S}, where \eqn{S} is the Yule coefficient
    in Legendre & Legendre.
}
}
