% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{yule}
\alias{yule}
\title{Yule distance}
\usage{
yule(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\description{
Yule distance
}
\details{
Relation to other definitions:
\itemize{
  \item Equivalent to yule() function in scipy.spatial.distance.
  \item Equivalent to 1 - S, where S is the Yule coefficient in Legendre & Legendre.
}
}
